+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 1.0.1
+* Fix 1 bullet per weapon
+* Fix keep weapons when stop work as cop
+* Keep skin and only change clothes when take or stop service (thanks Xtas3)
+
 ## 1.0.0
 * Original files added